requirements.txt
jupyter
transformers
langchain
Pillow
torch
pandas
huggingface_hub
text_generation
sentence_transformers
chromadb
streamlit